validation directory created, validation files renamed, test api moved to 5555
