+++ app/api.ts
@@ -1,8 +1,8 @@
 import axios from "axios";
 
-import { HealthRecordType } from "@/types/healthRecordValidationSchemas";
+import { HealthRecordType } from "@/validation/healthRecordSchema";
 
-const API_URL = "http://localhost:5000/health-records";
+const API_URL = "http://localhost:5555/health-records";
 
 export const getHealthRecord = async (id: number) => {
   try {
