+++ app/edit/EditStatus.tsx
@@ -5,9 +5,9 @@ import DropDownPicker from "react-native-dropdown-picker";
 import { SaveCancelButtons } from "@/components/formElements/SaveCancelButtons";
 import { LoadingIndicator } from "@/components/LoadingIndicator";
 import { useEditForm } from "@/hooks/useEditForm";
-import { StatusOptionsType } from "@/types/healthRecordValidationSchemas";
+import { StatusOptionsType } from "@/validation/healthRecordSchema";
 import { statusConfigs, statusOptions } from "@/utils/constants";
-import { validators } from "@/utils/validators";
+import { validators } from "@/validation/validators";
 
 const EditStatus = () => {
   const { localValue, setLocalValue, handleSave, loading } = useEditForm("status", validators.status);
