+++ package.json
@@ -4,7 +4,7 @@
   "version": "1.0.0",
   "scripts": {
     "start": "expo start",
-    "start:mock": "json-server --watch db.json --port 5000",
+    "start:mock": "json-server --watch db.json --port 5555",
     "reset-project": "node ./scripts/reset-project.js",
     "android": "expo start --android",
     "ios": "expo start --ios",
